Genii can be called
worked out the system to allow the creature to be called over, next up is working on creature stat increases. should probably work on genii instancing so that i can have base stats for baby creatures and a point to reset to during testing. also need to redo the inventory sys.